bin/legacy/empr-archive-0.3.1/cms/requirements.txt
asgiref==3.6.0
Django==4.1.7
Markdown==3.4.1
Pillow==9.4.0
sqlparse==0.4.4
